Rename new sitemap search components and routes to "page chooser search" to increase distinction between back-end advanced search
