+++ assets/cms/components/index.js
@@ -12,7 +12,7 @@ import ConcreteFileInput from './form/ConcreteFileInput'
 import ConcretePageInput from './form/ConcretePageInput'
 import ConcreteEventOccurrenceInput from './form/ConcreteEventOccurrenceInput'
 import ConcreteFileChooser from './file-manager/Chooser'
-import ConcreteSitemapSearch from './sitemap/SitemapSearch'
+import ConcretePageChooserSearch from './page/Chooser/ChooserSearch'
 import Board from './Board'
 import BoardSlot from './BoardSlot'
 
@@ -31,7 +31,7 @@ export default {
     ConcreteFileChooser,
     ConcreteFileInput,
     ConcretePageInput,
-    ConcreteSitemapSearch,
+    ConcretePageChooserSearch,
     ConcreteEventOccurrenceInput,
     Board,
     BoardSlot
